GenerateWorldMesh() in InitWorld now generates a world mesh of all tiles.
Moved some general graphics function to a new file called UtilityGraphics.h, for easy access.
