+++ Game/GameMain/Renderers/Renderers.cpp
@@ -2,26 +2,9 @@
 #include "../gameAI/AIComponents.h"
 #include "../gameUtility/UtilityComponents.h"
 #include "../gameUtility/UtilityEcsFunctions.h"
+#include "../gameUtility/UtilityGraphics.h"
 #include "../gameSceneObjects/SceneObjectComponents.h"
 #include "../gameWorld/OceanComponents.h"
-
-static inline uint32_t PACK(uint8_t c0, uint8_t c1, uint8_t c2, uint8_t c3) {
-	return (c0 << 24) | (c1 << 16) | (c2 << 8) | c3;
-}
-
-static const std::string GetShaderFilepath(const char* pFilename)
-{
-	std::string filepath = "..//";
-
-#ifdef _DEBUG
-	filepath.append("shaders_d//");
-#else
-	filepath.append("shaders//");
-#endif // _DEBUG
-	filepath.append(pFilename);
-
-	return filepath;
-}
 
 namespace ecs
 {
